requestProcess/requirements.txt
# Function dependencies, for example:
# package>=version
flask==1.0.2
slackclient==2.0.1
google-cloud-core==1.0.0
google-cloud-bigquery==1.14.0
googleapis-common-protos==1.6.0
google-cloud-storage==1.16.1
google-cloud-logging==1.11.0
google-cloud-pubsub==0.41.0